bot/requirements.txt
flask==3.0.*
requests==2.31.*
psycopg2-binary==2.9.*
